refactor: 이미지 파일 경로 수정
