+++ wecare-frontend/wecare-fe/src/screen/Home/HomeScreen.jsx
@@ -7,7 +7,7 @@ export default function HomeScreen() {
   return (
     <View testID="home-screen">
       <Text>Home Screen</Text>
-      <TimePicker />
+      {/* <TimePicker /> */}
     </View>
   );
 }
